fix(ac-ui): steady waterfall time axis + smooth fractional zoom
The waterfall/CWT Y-axis jumped around because the row-period estimator
was a 15%-EMA and every jitter pulse yanked the labels; zoom was also
coarse because rows_visible only lived as u32 so each scroll tick rounded
to a full-row jump.

Row period now tracks a median over the last 16 inter-arrival samples
with a 3% relative-change gate before publishing, so single stalls and
micro-jitter no longer flip labels. Rows_visible gains an f32 companion
that scroll updates smoothly; the u32 is round(f32) for the shader while
the time-axis labels read the f32 directly for continuous scaling.

+++ ac-rs/crates/ac-ui/src/render/grid.rs
@@ -6,9 +6,11 @@ use crate::theme;
 pub struct WaterfallTimeAxis {
     /// Seconds between successive waterfall rows (producer frame interval).
     pub row_period_s: f32,
-    /// Number of newest rows currently stretched across the cell (from
-    /// `CellView::rows_visible`). One label per ~4 subdivisions of the axis.
-    pub rows_visible: u32,
+    /// Fractional number of newest rows currently stretched across the cell.
+    /// Tracks `CellView::rows_visible_f` directly so the time axis grows
+    /// and shrinks smoothly while the shader-facing `u32` counterpart steps
+    /// at integer boundaries. One label per ~4 subdivisions of the axis.
+    pub rows_visible: f32,
 }
 
 pub fn draw_grid(
@@ -89,9 +91,9 @@ pub fn draw_grid(
             }
             let axis = time_axis.unwrap_or(WaterfallTimeAxis {
                 row_period_s: 0.1,
-                rows_visible: 256,
+                rows_visible: 256.0,
             });
-            let total_s = axis.rows_visible as f32 * axis.row_period_s;
+            let total_s = axis.rows_visible * axis.row_period_s;
             let ticks = time_ticks(total_s);
             for t_s in ticks {
                 let frac = if total_s > 0.0 { t_s / total_s } else { 0.0 };
